fix: keep detached workspace creation under file budget
